# [mypyc] Faster creation of a tuple from list/tuple