--- a/mypyc/irbuild/for_helpers.py
+++ b/mypyc/irbuild/for_helpers.py
@@ -86,9 +86,17 @@ def for_loop_helper(builder: IRBuilder, index: Lvalue, expr: Expression,
     builder.activate_block(exit_block)
 
 
-def for_loop_helper_simple(builder: IRBuilder, index: Lvalue, expr: Expression,
-                           body_insts: Callable[[Value], None], line: int) -> None:
+def for_loop_helper_with_index(builder: IRBuilder, index: Lvalue, expr: Expression,
+                               body_insts: Callable[[Value], None], line: int) -> None:
+    """Generate IR for a sequence iteration.
+       Compare to for_loop_helper, it would feed iteration index to body_insts.
 
+    Args:
+        index: the loop index Lvalue
+        expr: the expression to iterate over
+        body_insts: a function that generates the body of the loop.
+                    It needs a index as parameter.
+    """
     body_block, step_block, exit_block = BasicBlock(), BasicBlock(), BasicBlock()
     condition_block = BasicBlock()