### Description

mypyc: https://github.com/mypyc/mypyc/issues/769