@97littleleaf11 2021-03-16
I went through `translate_list_comprehension` and `comprehension_helper`. I would write similar code to fill in the blank tuple after I get `tuple_op`, is it correct?