@JukkaL 2021-03-16
> I went through translate_list_comprehension and comprehension_helper. I should write similar code to fill in the blank tuple after I get tuple_op, is it correct?

That's right! It's also possible that you can reuse `ForSequence` (or `ForEnumerate` / `ForRange`) to simplify the generation of the for loop (not sure though whether it's worth it).