@97littleleaf11 2021-03-16
It should work now, I will add some IR test tomorrow.