@97littleleaf11 2021-03-17
There is another implementation on a new personal branch: https://github.com/97littleleaf11/mypy/commit/4f0b92acb0401dc7ebc0051ef00ce9af3560c4fe