@TH3CHARLie 2021-03-18
`c_pyssize_t` has platform-specific size so we use `native_int` in IR output and it will be replaced with either `int32` or `int64` in test textual matching. So a quick fix is to replace every `int64` in your code with `native_int`. Later you can try to improve https://github.com/mypyc/mypyc/issues/776 to solve this some-what ugly workflow