@TH3CHARLie 2021-03-18
I kind of feel this issue is less good-first-issue. But that's good, you get to know more aspects about the codebase and workflow :D