@97littleleaf11 2021-03-18
```py
@benchmark
def tuple_build_test() -> None:
    a = []
    for i in range(1000):
        a.append(i)
    for i in range(1000):
        b = tuple(x + 1 for x in a)
```

new:

interpreted: 0.059748s (avg of 16 iterations; stdev 1.2%)
compiled:    0.017971s (avg of 16 iterations; stdev 3.6%)

compiled is 3.325x faster


master: 

interpreted: 0.058382s (avg of 17 iterations; stdev 0.96%)
compiled:    0.025852s (avg of 17 iterations; stdev 2.5%)

compiled is 2.258x faster