@TH3CHARLie 2021-03-18
you can use the resolve conversation button to make the PR page more concise